requirements.txt
fastapi==0.115.5
langchain_community==0.3.7
langchain_core==0.3.17
langchain_openai==0.2.8
pydantic==2.9.2
python-dotenv==1.0.1
Requests==2.32.3
tiktoken==0.8.0
uvicorn==0.32.0
python-multipart
pymupdf